+++ src/main/java/com/bergerkiller/bukkit/common/utils/NBTUtil.java
@@ -9,10 +9,10 @@ import com.bergerkiller.bukkit.common.nbt.CommonTagList;
 import com.bergerkiller.bukkit.common.nbt.NBTTagInfo;
 import com.bergerkiller.bukkit.common.reflection.classes.EntityLivingRef;
 import com.bergerkiller.bukkit.common.reflection.classes.NBTRef;
-import net.minecraft.server.v1_9_R1.*;
+import net.minecraft.server.v1_10_R1.*;
 import org.bukkit.block.BlockState;
-import org.bukkit.craftbukkit.v1_9_R1.inventory.CraftInventoryCustom;
-import org.bukkit.craftbukkit.v1_9_R1.inventory.CraftItemStack;
+import org.bukkit.craftbukkit.v1_10_R1.inventory.CraftInventoryCustom;
+import org.bukkit.craftbukkit.v1_10_R1.inventory.CraftItemStack;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.inventory.Inventory;
 
@@ -61,7 +61,7 @@ public class NBTUtil {
         if (nbtTagHandle == null) {
             return (byte) 0;
         }
-        return NBTRef.getTypeId.invoke(nbtTagHandle).byteValue();
+        return NBTRef.getTypeId.invoke(nbtTagHandle);
     }
 
     /**
